web_app/requirements.txt
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deep-translator==1.11.4
filelock==3.20.0
Flask==3.1.2
Flask-SocketIO==5.5.1
fsspec==2025.12.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.36.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
langdetect==1.0.9
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.2
numpy==2.3.5
packaging==25.0
python-aiml==0.9.3
python-engineio==4.12.3
python-socketio==5.15.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rfc3986==1.5.0
safetensors==0.7.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
wsproto==1.3.2
